pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reaper_mcp"
version = "0.1.0"
description = "A comprehensive MCP server for controlling REAPER with AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-reapy>=0.10.0",
    "mcp-server>=0.1.0",
    "numpy>=1.20.0",
    "librosa>=0.9.0",
]

[project.scripts]
reaper-mcp = "reaper_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.9.1",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
